+++ examples/resources/nxos_pim/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_pim.example "sys/pim"