+++ examples/resources/nxos_pim/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_pim" "example" {
+  admin_state = "enabled"
+}