+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Add nxos_ospf_area resource and data source
 - Add nxos_ospf_interface resource and data source
 - Add nxos_ospf_authentication resource and data source
+- Add nxos_pim resource and data source
 
 ## 0.2.3